Refactor tests by extracting projects and initializeProject functions
